Time: 665 ms (5.02%), Space: 49.5 MB (5.10%) - LeetHub

+++ 0907-sum-of-subarray-minimums/0907-sum-of-subarray-minimums.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int modulo=1e9+7;
     int sumSubarrayMins(vector<int>& arr) {
-        vector<int> nsl; // for distance between previous smaller element and current
+        vector<long long> nsl; // for distance between previous smaller element and current
         int n = arr.size();
-        vector<long long> nsr;  //for distance between next smaller eleement to right and current
+        vector<int> nsr;  //for distance between next smaller eleement to right and current
         stack<int> st;
         for(int i=0;i<arr.size();i++){
             if(st.empty())
